Ensure console is available when evaluating
The `eval` dependency has a bug in node v10 and above, where the
`console` won't be available to scripts, even if globals are supposed to
be available. See: pierrec/node-eval#18. For users of
static-site-generator-webpack-plugin, this means that index scripts
which contain logging statements will not work properly.

The bug is fixed in eval v0.1.4, so we just need to update the
dependency.

Test plan: `yarn test` still passes.